pyproject.toml
[tool.poetry]
name = "botvov"
version = "0.1.0"
description = "An AI assistant support answer questions and do tasks from user"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ollama = "^0.3.1"
qwen-agent = {path = "./Qwen-Agent/", extras = ["gui", "rag", "code-interpreter"], develop = true}
pydantic = "^2.8.2"
flask = "^3.0.3"
requests = "^2.32.3"
torch = "2.4.1"
soundfile = "0.12.1"
tqdm = "4.66.5"
numpy = "1.26.4"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
